requirements.txt
streamlit==1.32.2
pandas>=2.0
plotly>=6.0
matplotlib>=3.6
seaborn>=0.13
wordcloud>=1.9
huggingface-hub>=0.14.1   # used for InferenceApi
scikit-learn>=1.2
python-dotenv>=1.0
